pyproject.toml
[project]
name = "job-navigator"
version = "0.1.0"
description = "AI-powered job discovery app that matches users to the best roles based on their skills and preferences"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-deepseek>=0.1.4",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langsmith>=0.4.25",
    "streamlit>=1.49.1",
]
